requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==23.0.0
redis[async]==6.2.0
pydantic-settings>=2.10.1
pydantic==2.11.7
pyyaml==6.0.2
httpx==0.28.1